Average Marginal Effects

Provides the average marginal effects of a GLM model with bootstrapped confidence intervals.

frames(data, formula, family, ci_type = "perc", boot = 100, ci = 0.95)

Arguments

data
the data frame containing the variables for the model

formula
the model formula

family
the family of the glm (e.g., binomial, poisson)

ci_type
the type of boostrapped confidence interval; options are "perc", "basic", "bca"

boot
the number of bootstrapped samples; default is 100

ci
the confidence interval; the default is .975 which is the 95% confidence interval.

Details

Using the average marginal effects as discussed by Tamas Bartus (2005), the coefficients are transformed into probabilities (for binary outcomes) or remain in their original units (continuous outcomes).

References

Bartus, T. (2005). Estimation of marginal effects using margeff. The Stata Journal, 5(3), 309–329.