Marginal Mediation

Provides the ability to perform marginal mediation. Marginal mediation is particularly useful for situations where the mediator or outcome is categorical, a count, or some other non-normally distributed variable. The results provide the average marginal effects of the models, providing simple interpretation of the indirect effects.

mma(..., ind_effects, ci_type = "perc", boot = 500, ci = 0.95)

Arguments

...
the glm model objects; the first is the model with the outcome while the others are the mediated effects ("a" paths)

ind_effects
a vector of the desired indirect effects. Has the form "var1-var2".

ci_type
a string indicating the type of bootstrap method to use (currently "perc" and "basic" are available; "perc" is recommended). Further development will allow the Bias-Corrected bootstrap soon.

boot
the number of bootstrapped samples; default is 500

ci
the confidence interval; the default is .95 which is the 95% confidence interval.

Value

A list of class mma containing:

ind_effects

the indirect effects reported in the average marginal effect

dir_effects

the direct effects reported in the average marginal effect

ci_level

the confidence level

data

the original data frame

reported_ind

the indirect effects the user requested (in the ...)

boot

the number of bootstrap samples

model

the formulas of the individual sub-models

call

the original function call

Details

Using the average marginal effects as discussed by Tamas Bartus (2005), the coefficients are transformed into probabilities (for binary outcomes) or remain in their original units (continuous outcomes).

References

Bartus, T. (2005). Estimation of marginal effects using margeff. The Stata Journal, 5(3), 309–329.

MacKinnon, D. (2008). Introduction to Statistical Mediation Analysis. Taylor \& Francis, LLC.

Examples

## A minimal example:

library(furniture)
data(nhanes_2010)
bcpath = glm(marijuana ~ home_meals + gender + age + asthma,
           data = nhanes_2010,
           family = "binomial")
apath = glm(home_meals ~ gender + age + asthma,
           data = nhanes_2010,
           family = "gaussian")
(fit = mma(bcpath, apath,
           ind_effects = c("genderFemale-home_meals",
                           "age-home_meals",
                           "asthmaNo-home_meals"),
           boot = 10))

#> 
#> calculating a paths... b and c paths... Done.

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#>                                                                                  

#> ┌───────────────────────────────┐
#> │  Marginal Mediation Analysis  │
#> └───────────────────────────────┘
#> A marginal mediation model with:
#>    1 mediators
#>    3 indirect effects
#>    3 direct effects
#>    10 bootstrapped samples
#>    95% confidence interval
#>    n = 1417 
#> 
#> Formulas:
#>    ◌ marijuana ~ home_meals + gender + age + asthma
#>    ◌ home_meals ~ gender + age + asthma 
#> 
#> Regression Models: 
#> 
#>      marijuana ~ 
#>                           Est      SE   Est/SE P-Value
#>         (Intercept)  -0.39400 0.38028 -1.03608 0.30017
#>         home_meals   -0.04062 0.01363 -2.98051 0.00288
#>         genderFemale  0.43161 0.11723  3.68169 0.00023
#>         age           0.00276 0.01470  0.18754 0.85123
#>         asthmaNo     -0.00717 0.15004 -0.04778 0.96189
#> 
#>      home_meals ~ 
#>                           Est      SE   Est/SE P-Value
#>         (Intercept)   6.56883 0.76462  8.59100 0.00000
#>         genderFemale -1.34831 0.23910 -5.63913 0.00000
#>         age          -0.05689 0.03017 -1.88565 0.05955
#>         asthmaNo     -0.00428 0.31293 -0.01368 0.98909
#> 
#> Unstandardized Mediated Effects: 
#> 
#>    Indirect Effects: 
#> 
#>      marijuana ~ 
#>                                    Indirect    Lower   Upper
#>         genderFemale => home_meals  0.01312  0.00506 0.02342
#>         age => home_meals           0.00055  0.00000 0.00059
#>         asthmaNo => home_meals      0.00004 -0.00368 0.00042
#> 
#>    Direct Effects: 
#> 
#>      marijuana ~ 
#>                        Direct    Lower   Upper
#>         genderFemale  0.10430  0.06847 0.19614
#>         age           0.00066 -0.00478 0.00760
#>         asthmaNo     -0.00172 -0.09373 0.07559
#>